requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.19.2
pandas==1.1.3
streamlit==1.2.0
streamlit_drawable_canvas==0.8.0
Pillow==8.0.1
tensorflow==2.7.0
scikit-learn==1.0